# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"misnote API","version":"0.1.0"},"paths":{"/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","operationId":"get_me_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/subjects":{"get":{"tags":["subjects"],"summary":"List Subjects","operationId":"list_subjects_v1_subjects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SubjectResponse"},"type":"array","title":"Response List Subjects V1 Subjects Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["subjects"],"summary":"Create Subject","operationId":"create_subject_v1_subjects_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/subjects/{subject_id}":{"put":{"tags":["subjects"],"summary":"Update Subject","operationId":"update_subject_v1_subjects__subject_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"subject_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Subject Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["subjects"],"summary":"Delete Subject","operationId":"delete_subject_v1_subjects__subject_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"subject_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Subject Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/subjects/{subject_id}/units":{"get":{"tags":["units"],"summary":"List Units","operationId":"list_units_v1_subjects__subject_id__units_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"subject_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Subject Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UnitResponse"},"title":"Response List Units V1 Subjects  Subject Id  Units Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["units"],"summary":"Create Unit","operationId":"create_unit_v1_subjects__subject_id__units_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"subject_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Subject Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnitCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/units/{unit_id}":{"put":{"tags":["units"],"summary":"Update Unit","operationId":"update_unit_v1_units__unit_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"unit_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Unit Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnitUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["units"],"summary":"Delete Unit","operationId":"delete_unit_v1_units__unit_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"unit_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Unit Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/questions":{"get":{"tags":["questions"],"summary":"List Questions","operationId":"list_questions_v1_questions_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"subject_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Subject Id"}},{"name":"unit_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Unit Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionResponse"},"title":"Response List Questions V1 Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["questions"],"summary":"Create Question","operationId":"create_question_v1_questions_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/questions/{question_id}":{"get":{"tags":["questions"],"summary":"Get Question","operationId":"get_question_v1_questions__question_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["questions"],"summary":"Update Question","operationId":"update_question_v1_questions__question_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["questions"],"summary":"Delete Question","operationId":"delete_question_v1_questions__question_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/mistake-notes/today":{"get":{"tags":["mistake-notes"],"summary":"List Today","operationId":"list_today_v1_mistake_notes_today_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MistakeNoteResponse"},"type":"array","title":"Response List Today V1 Mistake Notes Today Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/mistake-notes/mastered":{"get":{"tags":["mistake-notes"],"summary":"List Mastered","operationId":"list_mastered_v1_mistake_notes_mastered_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MistakeNoteResponse"},"title":"Response List Mastered V1 Mistake Notes Mastered Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/mistake-notes":{"get":{"tags":["mistake-notes"],"summary":"List Active","operationId":"list_active_v1_mistake_notes_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MistakeNoteResponse"},"title":"Response List Active V1 Mistake Notes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/mistake-notes/{note_id}":{"get":{"tags":["mistake-notes"],"summary":"Get Note","operationId":"get_note_v1_mistake_notes__note_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"note_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Note Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MistakeNoteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["mistake-notes"],"summary":"Update Note","operationId":"update_note_v1_mistake_notes__note_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"note_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Note Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MistakeNoteUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MistakeNoteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/mistake-notes/{note_id}/status":{"put":{"tags":["mistake-notes"],"summary":"Update Status","operationId":"update_status_v1_mistake_notes__note_id__status_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"note_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Note Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MistakeNoteStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MistakeNoteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/questions/{question_id}/attempts":{"post":{"tags":["attempts"],"summary":"Create Attempt","operationId":"create_attempt_v1_questions__question_id__attempts_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttemptCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttemptResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["attempts"],"summary":"List Attempts","operationId":"list_attempts_v1_questions__question_id__attempts_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttemptHistoryItem"},"title":"Response List Attempts V1 Questions  Question Id  Attempts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"AttemptCreate":{"properties":{"user_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Answer"},"is_correct":{"type":"boolean","title":"Is Correct"}},"type":"object","required":["is_correct"],"title":"AttemptCreate"},"AttemptHistoryItem":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"question_id":{"type":"string","format":"uuid","title":"Question Id"},"user_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Answer"},"is_correct":{"type":"boolean","title":"Is Correct"},"answered_at":{"type":"string","format":"date-time","title":"Answered At"}},"type":"object","required":["id","question_id","user_answer","is_correct","answered_at"],"title":"AttemptHistoryItem"},"AttemptResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"question_id":{"type":"string","format":"uuid","title":"Question Id"},"user_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Answer"},"is_correct":{"type":"boolean","title":"Is Correct"},"answered_at":{"type":"string","format":"date-time","title":"Answered At"},"mistake_note_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Mistake Note Id"},"correct_streak":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Correct Streak"},"mastery_suggested":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Mastery Suggested"}},"type":"object","required":["id","question_id","user_answer","is_correct","answered_at","mistake_note_id","correct_streak","mastery_suggested"],"title":"AttemptResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MistakeNoteResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"question":{"$ref":"#/components/schemas/QuestionRef"},"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"},"learning":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Learning"},"status":{"type":"string","enum":["active","mastered"],"title":"Status"},"wrong_count":{"type":"integer","title":"Wrong Count"},"correct_streak":{"type":"integer","title":"Correct Streak"},"next_review_at":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Next Review At"}},"type":"object","required":["id","question","memo","learning","status","wrong_count","correct_streak","next_review_at"],"title":"MistakeNoteResponse"},"MistakeNoteStatusUpdate":{"properties":{"status":{"type":"string","enum":["active","mastered"],"title":"Status"}},"type":"object","required":["status"],"title":"MistakeNoteStatusUpdate"},"MistakeNoteUpdate":{"properties":{"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"},"learning":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Learning"},"next_review_at":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Next Review At"}},"type":"object","title":"MistakeNoteUpdate"},"QuestionCreate":{"properties":{"subject_id":{"type":"string","format":"uuid","title":"Subject Id"},"unit_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Unit Id"},"question_text":{"type":"string","title":"Question Text"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"},"memo":{"type":"string","minLength":1,"title":"Memo"},"learning":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Learning"},"next_review_at":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Next Review At"}},"type":"object","required":["subject_id","question_text","memo"],"title":"QuestionCreate"},"QuestionRef":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"subject":{"$ref":"#/components/schemas/SubjectRef"},"unit":{"anyOf":[{"$ref":"#/components/schemas/UnitRef"},{"type":"null"}]},"question_text":{"type":"string","title":"Question Text"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"}},"type":"object","required":["id","subject","unit","question_text","correct_answer"],"title":"QuestionRef"},"QuestionResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"subject":{"$ref":"#/components/schemas/SubjectRef"},"unit":{"anyOf":[{"$ref":"#/components/schemas/UnitRef"},{"type":"null"}]},"question_text":{"type":"string","title":"Question Text"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"mistake_note_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Mistake Note Id"}},"type":"object","required":["id","subject","unit","question_text","correct_answer","created_at","mistake_note_id"],"title":"QuestionResponse"},"QuestionUpdate":{"properties":{"subject_id":{"type":"string","format":"uuid","title":"Subject Id"},"unit_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Unit Id"},"question_text":{"type":"string","title":"Question Text"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"}},"type":"object","required":["subject_id","question_text"],"title":"QuestionUpdate"},"SubjectCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"SubjectCreate"},"SubjectRef":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"SubjectRef"},"SubjectResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"SubjectResponse"},"SubjectUpdate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"SubjectUpdate"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UnitCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"UnitCreate"},"UnitRef":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"UnitRef"},"UnitResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"subject_id":{"type":"string","format":"uuid","title":"Subject Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","subject_id","name"],"title":"UnitResponse"},"UnitUpdate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"UnitUpdate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password"},"name":{"type":"string","minLength":1,"title":"Name"}},"type":"object","required":["email","password","name"],"title":"UserRegister"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","email","name"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}